# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QR Menu System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /languages:
    get:
      operationId: listLanguages
      tags: [languages]
      summary: List all languages
      responses:
        "200":
          description: List of languages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Language"
    post:
      operationId: createLanguage
      tags: [languages]
      summary: Create a language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageInput"
      responses:
        "201":
          description: Created language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Language"
  /languages/{id}:
    patch:
      operationId: updateLanguage
      tags: [languages]
      summary: Update a language
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageUpdate"
      responses:
        "200":
          description: Updated language
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Language"
    delete:
      operationId: deleteLanguage
      tags: [languages]
      summary: Delete a language
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Categories
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of categories with translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryWithTranslations"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithTranslations"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category with translations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithTranslations"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithTranslations"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /categories/reorder:
    post:
      operationId: reorderCategories
      tags: [categories]
      summary: Reorder categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered

  # Products
...
  /analytics/views:
    get:
      operationId: getViewTimeseries
      tags: [analytics]
      summary: Get view timeseries data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
      responses:
        "200":
          description: Timeseries data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeseriesPoint"
  /analytics/top-products:
    get:
      operationId: getTopProducts
      tags: [analytics]
      summary: Get top viewed products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
  /analytics/lang-timeseries:
    get:
      operationId: getLangTimeseries
      tags: [analytics]
      summary: Get daily view timeseries broken down by language
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: Language-specific timeseries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LangTimeseriesPoint"

  # AI
  /ai/generate:
    post:
      operationId: generateAiContent
      tags: [ai]
      summary: Generate AI content suggestions for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateInput"
      responses:
        "200":
          description: AI-generated suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerateResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Object Storage
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Language:
      type: object
      required: [id, code, name, isDefault, isActive, sortOrder]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
    LanguageInput:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
    LanguageUpdate:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
    CategoryTranslation:
      type: object
      required: [id, categoryId, languageCode, name]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
    CategoryTranslationInput:
      type: object
      required: [languageCode, name]
      properties:
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
    CategoryWithTranslations:
      type: object
      required: [id, sortOrder, isActive, translations]
      properties:
        id:
          type: integer
        imagePath:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        translations:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTranslation"
    CategoryInput:
      type: object
      required: [translations]
      properties:
        imagePath:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        translations:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTranslationInput"
    CategoryUpdate:
      type: object
      properties:
        imagePath:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        translations:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTranslationInput"
...
    ProductNutrition:
      type: object
      properties:
        calories:
          type: ["number", "null"]
        energy:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        isAiGenerated:
          type: boolean
...
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
...
    TimeseriesPoint:
      type: object
      required: [date, menuViews, productViews]
      properties:
        date:
          type: string
        menuViews:
          type: integer
        productViews:
          type: integer
    TopProduct:
      type: object
      required: [productId, viewCount, name]
      properties:
        productId:
          type: integer
        viewCount:
          type: integer
        name:
          type: string
        imagePath:
          type: ["string", "null"]
...
    AiGenerateInput:
      type: object
      required: [productName]
      properties:
        productName:
          type: string
        notes:
          type: ["string", "null"]
        languages:
          type: array
          items:
            type: string
    AiGenerateResult:
      type: object
      properties:
        translations:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AiTranslation"
        allergens:
          type: array
          items:
            type: string
        nutrition:
          $ref: "#/components/schemas/ProductNutrition"
    AiTranslation:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        allergenNote:
          type: ["string", "null"]
        specialNote:
          type: ["string", "null"]
...
    LangTimeseriesPoint:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
        tr:
          type: integer
        en:
          type: integer
        ru:
          type: integer
        ar:
          type: integer